--- a/lib/elixir/test/elixir/version_test.exs
+++ b/lib/elixir/test/elixir/version_test.exs
@@ -49,16 +49,30 @@ defmodule VersionTest do
 
   test "lexes specifications properly" do
     assert Parser.lexer("== != > >= < <= ~>", []) == [:==, :!=, :>, :>=, :<, :<=, :~>]
-    assert Parser.lexer("2.3.0", []) == [:==, "2.3.0"]
-    assert Parser.lexer("!2.3.0", []) == [:!=, "2.3.0"]
+    assert Parser.lexer("2.3.0", []) == [:==, {2, 3, 0, [], []}]
+    assert Parser.lexer("!2.3.0", []) == [:!=, {2, 3, 0, [], []}]
     assert Parser.lexer(">>=", []) == [:>, :>=]
-    assert Parser.lexer(">2.4.0", []) == [:>, "2.4.0"]
-    assert Parser.lexer("> 2.4.0", []) == [:>, "2.4.0"]
-    assert Parser.lexer("    >     2.4.0", []) == [:>, "2.4.0"]
-    assert Parser.lexer(" or 2.1.0", []) == [:||, :==, "2.1.0"]
-    assert Parser.lexer(" and 2.1.0", []) == [:&&, :==, "2.1.0"]
-    assert Parser.lexer(">= 2.0.0 and < 2.1.0", []) == [:>=, "2.0.0", :&&, :<, "2.1.0"]
-    assert Parser.lexer(">= 2.0.0 or < 2.1.0", []) == [:>=, "2.0.0", :||, :<, "2.1.0"]
+    assert Parser.lexer(">2.4.0", []) == [:>, {2, 4, 0, [], []}]
+    assert Parser.lexer("> 2.4.0", []) == [:>, {2, 4, 0, [], []}]
+    assert Parser.lexer("    >     2.4.0", []) == [:>, {2, 4, 0, [], []}]
+    assert Parser.lexer(" or 2.1.0", []) == [:||, :==, {2, 1, 0, [], []}]
+    assert Parser.lexer(" and 2.1.0", []) == [:&&, :==, {2, 1, 0, [], []}]
+
+    assert Parser.lexer(">= 2.0.0 and < 2.1.0", []) == [
+             :>=,
+             {2, 0, 0, [], []},
+             :&&,
+             :<,
+             {2, 1, 0, [], []}
+           ]
+
+    assert Parser.lexer(">= 2.0.0 or < 2.1.0", []) == [
+             :>=,
+             {2, 0, 0, [], []},
+             :||,
+             :<,
+             {2, 1, 0, [], []}
+           ]
   end
 
   test "parse/1" do
@@ -212,6 +226,12 @@ defmodule VersionTest do
       refute Version.match?("0.10.0", "~> 0.9.3-dev")
 
       refute Version.match?("0.3.0-dev", "~> 0.2.0")
+
+      assert Version.match?("1.11.0-dev", "~> 1.11-dev")
+      assert Version.match?("1.11.0", "~> 1.11-dev")
+      assert Version.match?("1.12.0", "~> 1.11-dev")
+      refute Version.match?("1.10.0", "~> 1.11-dev")
+      refute Version.match?("2.0.0", "~> 1.11-dev")
 
       assert_raise Version.InvalidRequirementError, fn ->
         Version.match?("3.0.0", "~> 3")