@michalmuskala 2020-02-11
AFAIK with compiled match specs, it would also be possible to evaluate multiple versions against single requirement in one go, or even filter only matching requirements.

For example:
```elixir
ms = :ets.fun2ms(fn {x, y, z} = value when x > 1 and y < 2 and z > 0 -> value end)
cms = :ets.match_spec_compile(ms)
:ets.match_spec_run([{2, 1, 1}, {2, 0, 0}, {2, 1, 2}, {1, 0, 0}], ms)
#=> [{2, 1, 1}, {2, 1, 2}]
```

This "batch" mode filtering could prove more efficient than evaluating candidates one by one (I didn't do any measurements).