@ericmj 2020-02-11
Thanks for pointing this out, I ran some quick measurements with the benchmarks we have.

When chunking versions of 5 and passing the chunk to `:ets.match_spec_run/2` we get a 17% speed improvement over singles, when chunking 20 versions we get a 30% speed improvement. It doesn't seem likely we will get close to the 500% improvement we get from traditional pattern matches.

Is there anything suggesting that matching lists should have significant speed ups over matching single elements?